move parseFullXMLDocs to SourceDeclaration.swift
