+++ Source/SourceKittenFramework/File.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import SwiftXPC
-import SWXMLHash
 
 /// Represents a source file.
 public struct File {
@@ -350,57 +349,4 @@ private func isDeclarationOrCommentMark(dictionary: XPCDictionary) -> Bool {
             (kind == SyntaxKind.CommentMark.rawValue || SwiftDeclarationKind(rawValue: kind) != nil)
     }
     return false
-}
-
-/**
-Parse XML from `key.doc.full_as_xml` from `cursor.info` request.
-
-- parameter xmlDocs: Contents of `key.doc.full_as_xml` from SourceKit.
-
-- returns: XML parsed as an `XPCDictionary`.
-*/
-public func parseFullXMLDocs(xmlDocs: String) -> XPCDictionary? {
-    let cleanXMLDocs = xmlDocs.stringByReplacingOccurrencesOfString("<rawHTML>",  withString: "")
-                              .stringByReplacingOccurrencesOfString("</rawHTML>", withString: "")
-    return SWXMLHash.parse(cleanXMLDocs).children.first.map { rootXML in
-        var docs = XPCDictionary()
-        docs[SwiftDocKey.DocType.rawValue] = rootXML.element?.name
-        docs[SwiftDocKey.DocFile.rawValue] = rootXML.element?.attributes["file"]
-        docs[SwiftDocKey.DocLine.rawValue] = rootXML.element?.attributes["line"].map {
-            Int64(($0 as NSString).integerValue)
-        }
-        docs[SwiftDocKey.DocColumn.rawValue] = rootXML.element?.attributes["column"].map {
-            Int64(($0 as NSString).integerValue)
-        }
-        docs[SwiftDocKey.DocName.rawValue] = rootXML["Name"].element?.text
-        docs[SwiftDocKey.DocUSR.rawValue] = rootXML["USR"].element?.text
-        docs[SwiftDocKey.DocDeclaration.rawValue] = rootXML["Declaration"].element?.text
-        let parameters = rootXML["Parameters"].children
-        if parameters.count > 0 {
-            docs[SwiftDocKey.DocParameters.rawValue] = parameters.map {
-                return [
-                    "name": $0["Name"].element?.text ?? "",
-                    "discussion": childrenAsArray($0["Discussion"]) ?? []
-                ] as XPCDictionary
-            } as XPCArray
-        }
-        docs[SwiftDocKey.DocDiscussion.rawValue] = childrenAsArray(rootXML["Discussion"])
-        docs[SwiftDocKey.DocResultDiscussion.rawValue] = childrenAsArray(rootXML["ResultDiscussion"])
-        return docs
-    }
-}
-
-/**
-Returns an `XPCArray` of `XPCDictionary` items from `indexer` children, if any.
-
-- parameter indexer: `XMLIndexer` to traverse.
-*/
-private func childrenAsArray(indexer: XMLIndexer) -> XPCArray? {
-    let children = indexer.children
-    if children.count > 0 {
-        return children.flatMap({ $0.element }).map {
-            [$0.name: $0.text ?? ""] as XPCDictionary
-        } as XPCArray
-    }
-    return nil
 }
